Fixed minor backend issue as well. Router now added, splash page done, lazy loading added to the second half as it was getting bulky
