+++ frontend/package.json
@@ -21,6 +21,8 @@
     "react-d3-cloud": "^1.0.6",
     "react-dom": "^18.2.0",
     "react-fade-in": "^2.0.1",
+    "react-icons": "^4.8.0",
+    "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
     "react-spinners": "^0.13.8",
     "react-tagcloud": "^2.3.1",
