+++ frontend/src/css/App.css
@@ -1,5 +1,3 @@
-@import url('tweet.css');
-
 .logo {
   border-radius: 50%;
   display: inline-block;
